chunk log filtering into windows
